backend/requirements.txt
dill>=0.4.0
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart==0.0.18
python-jose[cryptography]>=3.4.0
passlib==1.7.4
httpx
email-validator==2.0.0
python-dotenv>=1.0.1
bcrypt==4.0.1

# Development tools
black>=23.0.0

# Azure Services
azure-storage-blob==12.25.1
azure-identity==1.23.0
openai

# Databases
motor
psycopg2-binary

# Document processing
docling
langchain-text-splitters
beautifulsoup4
grobid-client-python

# Vector database
pymilvus==2.5.11
pymilvus.model==0.3.2

# Additional utilities
aiofiles>=24.1.0

# Agent framework dependencies
langgraph
langchain
langchain-core==0.3.81
langchain-google-genai
google-genai
google-generativeai
wikipedia-api>=0.8.1
wikipedia==1.4.0

